requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
gunicorn
python-dotenv
psycopg2-binary
pdfkit 